refactor(login):split model classes into indvidual files
