+++ lib/modules/login/models/login_request.dart
@@ -1,0 +1,16 @@
+class LoginRequest {
+  final String username;
+  final String password;
+
+  LoginRequest({
+    required this.username,
+    required this.password,
+  });
+
+  Map<String, dynamic> toJson() {
+    return {
+      'username': username,
+      'password': password,
+    };
+  }
+}
