+++ lib/modules/login/models/login_response.dart
@@ -1,20 +1,3 @@
-class LoginRequest {
-  final String username;
-  final String password;
-
-  LoginRequest({
-    required this.username,
-    required this.password,
-  });
-
-  Map<String, dynamic> toJson() {
-    return {
-      'username': username,
-      'password': password,
-    };
-  }
-}
-
 class LoginResponse {
   final String token;
   final bool success;
