+++ lib/modules/login/services/login_mock_service.dart
@@ -1,4 +1,4 @@
-import '../models/login_model.dart';
+import '../models/login_response.dart';
 
 class LoginMockService {
   Future<LoginResponse> login(String username, String password) async {
